Addressing code review feedback: - Splitting tests into  different files and only running relavent tests - Replacing timeout and function with Eventually block [#145450903]
