+++ src/integration-tests/workload/iaas_lbs/lb_suite_test.go
@@ -9,24 +9,21 @@ import (
 	. "github.com/onsi/gomega"
 )
 
-func TestGcpLb(t *testing.T) {
+func TestIAASLb(t *testing.T) {
 	RegisterFailHandler(Fail)
-	RunSpecs(t, "GcpLb Suite")
+	RunSpecs(t, "IAASLb Suite")
 }
 
 var (
 	runner        *test_helpers.KubectlRunner
 	nginxSpec     = test_helpers.PathFromRoot("specs/nginx.yml")
-	nginxLBSpec   = test_helpers.PathFromRoot("specs/nginx-lb.yml")
 	workerAddress string
 	nodePort      string
-	iaas          string
 )
 
 var _ = BeforeSuite(func() {
 	workerAddress = os.Getenv("WORKLOAD_ADDRESS")
 	nodePort = os.Getenv("WORKLOAD_PORT")
-	iaas = os.Getenv("IAAS")
 	runner = test_helpers.NewKubectlRunner()
 	runner.RunKubectlCommand("create", "namespace", runner.Namespace()).Wait("60s")
 })
